refactor: :hammer: move `CompanyConstants` to `Constants`
